Move fasttest into 3rdparty folder
