+++ 3rdparty/fasttext/build.gradle
@@ -14,5 +14,20 @@ dependencies {
     testImplementation("org.testng:testng:${testng_version}") {
         exclude group: "junit", module: "junit"
     }
+    testImplementation project(":mxnet:mxnet-engine")
+    testImplementation "ai.djl.mxnet:mxnet-native-auto:${mxnet_version}"
     testImplementation "org.slf4j:slf4j-simple:${slf4j_version}"
+}
+
+publishing {
+    publications {
+        maven(MavenPublication) {
+            pom {
+                artifactId "fasttext-engine"
+                name = "Fasttext Engine Adapter"
+                description = "Fasttext Engine Adapter for DJL"
+                url = "http://www.djl.ai/fasttext/${project.name}"
+            }
+        }
+    }
 }
