+++ 3rdparty/fasttext/src/main/java/ai/djl/fasttext/zoo/package-info.java
