+++ 3rdparty/fasttext/src/main/javadoc/overview.html
@@ -1,0 +1,13 @@
+<html>
+<head>
+  <meta charset="UTF-8">
+</head>
+<body>
+<p>This document is the API specification for the Deep Java Library (DJL) NLP support with fastText.</p>
+
+<p>
+  The fastText module contains the NLP suport with fastText.
+</p>
+
+</body>
+</html>
