+++ 3rdparty/fasttext/src/main/resources/META-INF/services/ai.djl.repository.zoo.ZooProvider
