+++ 3rdparty/fasttext/src/test/java/ai/djl/fasttext/package-info.java
