+++ 3rdparty/fasttext/src/test/resources/mlrepo/dataset/nlp/ai/djl/fasttext/cooking_stackexchange/metadata.json
