+++ 3rdparty/fasttext/src/test/resources/mlrepo/model/nlp/textclassification/ai/djl/fasttext/cooking_stackexchange/metadata.json
