--- fasttext/fasttext-engine/build.gradle
@@ -1,30 +1,0 @@
-group "ai.djl.fasttext"
-
-repositories {
-    maven {
-        url 'https://oss.sonatype.org/content/repositories/snapshots/'
-    }
-}
-
-dependencies {
-    api "com.github.vinhkhuc:jfasttext:0.4"
-    api project(":api")
-    api project(":basicdataset")
-
-    testImplementation("org.testng:testng:${testng_version}") {
-        exclude group: "junit", module: "junit"
-    }
-    testImplementation "org.slf4j:slf4j-simple:${slf4j_version}"
-}
-
-publishing {
-    publications {
-        maven(MavenPublication) {
-            pom {
-                name = "Fasttext Engine Adapter"
-                description = "Fasttext Engine Adapter for DJL"
-                url = "http://www.djl.ai/fasttext/${project.name}"
-            }
-        }
-    }
-}
