--- fasttext/fasttext-engine/gradlew
@@ -1,1 +1,0 @@
-../../gradlew
