--- fasttext/fasttext-engine/src/main/java/ai/djl/fasttext/dataset/FtDataset.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
- * with the License. A copy of the License is located at
- *
- * http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
- * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
-package ai.djl.fasttext.dataset;
-
-import ai.djl.training.dataset.Dataset;
-import java.io.IOException;
-import java.nio.file.Path;
-
-/** An interface loads fastText dataset. */
-public interface FtDataset extends Dataset {
-
-    /**
-     * Returns cached fastText dataset file path.
-     *
-     * @return cached fastText dataset file path
-     * @throws IOException when IO operation fails in loading a resource
-     */
-    Path getInputFile() throws IOException;
-}
