--- fasttext/fasttext-engine/src/main/java/ai/djl/fasttext/dataset/FtDatasets.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
- * with the License. A copy of the License is located at
- *
- * http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
- * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
-package ai.djl.fasttext.dataset;
-
-import ai.djl.repository.Repository;
-
-/** An interface which contains datasets for fastText that are hosted on https://mlrepo.djl.ai/. */
-public interface FtDatasets {
-
-    String DJL_REPO_URL = "https://mlrepo.djl.ai/";
-
-    Repository REPOSITORY = Repository.newInstance("FtDataset", DJL_REPO_URL);
-
-    String GROUP_ID = "ai.djl.fasttext";
-}
