--- fasttext/fasttext-engine/src/main/javadoc/overview.html
@@ -1,14 +1,0 @@
-<html>
-<head>
-  <meta charset="UTF-8">
-</head>
-<body>
-<p>This document is the API specification for the Deep Java Library (DJL) fastText Engine.</p>
-
-<p>
-  The fastText Engine module contains the fastText implementation of the DJL EngineProvider.
-  See <a href="https://github.com/awslabs/djl/tree/master/fasttext/fasttext-engine">here</a> for more details.
-</p>
-
-</body>
-</html>
