--- fasttext/fasttext-engine/src/main/resources/META-INF/services/ai.djl.engine.EngineProvider
@@ -1,1 +1,0 @@
-ai.djl.fasttext.engine.FtEngineProvider
