+++ settings.gradle
@@ -15,7 +15,6 @@ include ':tensorflow:tensorflow-native-auto'
 include ':pytorch:pytorch-engine'
 include ':pytorch:pytorch-model-zoo'
 include ':pytorch:pytorch-native'
-include ':fasttext:fasttext-engine'
 include ':onnxruntime:onnxruntime-engine'
 include ':testing'
 include ':3rdparty:aws-ai'
